fix uri query decoding
If we decode the query as a uri component, we may end up with an invalid query
string. For example:

    http://example.com/?a=%3d

Is a uri with a query string whose value is "?a=%3d". If decoding the query
string as a uri component, its decoded value will be "?a==" which is not a
valid query string.

The proper way to decode is to parse query components. For this example, we
end up with one query parameter with the name "a" and the value "=".

This way, we also are sure to correctly encode uris.